Keep state path validation portable

+++ src/session_store.rs
@@ -1296,7 +1296,7 @@ fn validate_state_ancestor(
         );
     }
     let mode = metadata.mode();
-    let protected_by_sticky_root = owner == 0 && mode & libc::S_ISVTX as u32 != 0;
+    let protected_by_sticky_root = owner == 0 && mode & libc::S_ISVTX != 0;
     if mode & 0o022 != 0
         && !protected_by_sticky_root
         && !(is_immediate_parent && owner == effective_uid)
